LPS-124006 Extract ImageSelectorSize
