+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/Plugin.scss
@@ -1,5 +1,4 @@
 @import './item-configuration-panels/FieldSet';
-@import './item-configuration-panels/ImagePropertiesPanel';
 @import './item-configuration-panels/MappingPanel';
 @import './item-configuration-panels/RowStylesPanel';
