--- modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/item-configuration-panels/ImagePropertiesPanel.scss
@@ -1,6 +1,0 @@
-@import 'atlas-variables';
-
-.page-editor__image-properties-panel__resolution-label {
-	color: $secondary;
-	font-size: 0.75rem;
-}
